rm-bg/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0,<5.0.0
pillow>=8.0.0,<12.0
scikit-image